Start to split up test-pack by bundle entry
