+++ tool/site/client/test-pack/helpers.ts
@@ -1,0 +1,17 @@
+export interface StringKeyedObject {
+  [k: string]: string
+}
+
+export function groupBy<TKey, TValue>(list: Array<TValue>, keyGetter: (item: TValue) => TKey): Map<TKey, Array<TValue>> {
+  const map = new Map()
+  list.forEach(item => {
+    const key = keyGetter(item)
+    const collection = map.get(key)
+    if (!collection) {
+      map.set(key, [item])
+    } else {
+      collection.push(item)
+    }
+  })
+  return map
+}
